Adds Game States
Added a gamestate singleton that is responsible for holding the gamestate
and knowing what actions can be taken next, who the current player is,
and transitions between game states.

For now it knows and does nothing with card logic or game rule logic.
This may be added with some sort of callback system in the future, but
wanted to keep it simple and with one clear responsibility.